Modify Subscription ID generation & RelayPool.subscribe.
The subscription IDs generated when invoking RelayPool.subscribe
used to be a 32-byte hex string. Some relays were rejecting this
as too long so they have been shortened to an 8-byte hex string.

RelayPool.subscribe is now an asynchronous operation returning a
Future<String>. This future completes when all expected NIP-15
EOSE notices have been received and returns the subscription's
ID. This gives a way for clients to know that the initial
retreival of requested events has completed and any future events
will be new events.